// agent/src/app/shared/model/ana-chat.models.ts
export interface ANAClientInfo {
	id: string;
	medium: number;
}

export interface ANAMeta {
	id: string;
	sender: ANAClientInfo;
	recipient: ANAClientInfo;
	senderType: SenderType;
	timestamp: number;
	sessionId: string;
	responseTo: string;
	flowId: string;
	previousFlowId: string;
	currentFlowId: string;
}

export enum InputType {
	LOCATION = 7, //Click, Complex
	DATE = 5, //Click, Complex
	TIME = 6, //Click, Complex
	ADDRESS = 4, //Click, Complex
	MEDIA = 8, //Click, Non Complex
	OPTIONS = 10, //Click, Non Complex
	LIST = 9, //Click, Complex
	PHONE = 3,
	EMAIL = 2,
	NUMERIC = 1,
	TEXT = 0,
}

export enum SenderType {
	AGENT = 3,
	ANA = 1,
	AI = 2,
	USER = 0
}

export enum MediaType {
	IMAGE = 0,
	VIDEO = 2,
	FILE = 3,
	AUDIO = 1
}

export enum ButtonType {
	ACTION = 2,
	QUICK_REPLY = 1,
	URL = 0
}

export enum MessageType {
	CAROUSEL = 1,
	INPUT = 2,
	EXTERNAL = 3,
	SIMPLE = 0
}
export enum EventType {
	DISCONNECTED = 0,
	CONNECTED = 1,
	IDLE = 2,
	SESSION_STARTED = 3,
	SESSION_CLOSED = 4,
	CHAT_ALLOCATION = 5,
	CHAT_DEALLOCATION = 6,
	MESSAGE_COUNT = 7,
	RESPONSE_NEEDED = 8,
	ACTIVE = 9,
	INACTIVE = 10,
	TYPING = 11,
	AWAY = 12,
	ACK = 13,
	NO_AGENT_FOUND = 14,
	INTENT_TO_HANDOVER = 15,
	INTERACTION = 16,
	QUEUED = 17,
	PRIORITY_QUEUED = 18,
	ALLOCATION_REQUEST = 19,
	OUT_OF_BUSINESS_HOURS = 20,
	SET_SESSION_DATA = 21,
	HANDOVER = 22
}
export enum Bool {
	TRUE = 1,
	FALSE = 0,
}

export interface BaseContent {
	mandatory?: Bool;
}

export interface SimpleMessageData extends ANAMessageData {
	content: SimpleContent;
}

export interface CarouselMessageData extends ANAMessageData {
	content: CarouselContent;
}

export interface InputMessageData extends ANAMessageData {
	content: InputContent;
}

export interface ANAMessageData {
	type: MessageType;
	content: BaseContent;
}
// ...
export interface ANAEvent {
	type: EventType;
	data?: any;
	channel?: string;
}

export interface ANAChatMessage {
	meta: ANAMeta;
	data: ANAMessageData;
	events?: ANAEvent[];
	customData?: CustomData;
}

export interface CustomData {
	ackId?: string;
	status?: 'pending' | 'sent' | 'delivered';
}